# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://api.revmetrix.io"
    }
  ],
  "paths": {
    "/api/Insert/Insert": {
      "post": {
        "tags": [
          "Insert"
        ],
        "operationId": "Insert",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Research/Testing": {
      "post": {
        "tags": [
          "Research"
        ],
        "operationId": "Testing",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/ByeBye": {
      "post": {
        "tags": [
          "SuperUser"
        ],
        "operationId": "ByeBye",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/HashAndSalt": {
      "post": {
        "tags": [
          "SuperUser"
        ],
        "operationId": "HashAndSalt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              }
            }
          }
        }
      }
    },
    "/api/Test/Test": {
      "get": {
        "tags": [
          "Test"
        ],
        "operationId": "Test",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Test/TestAuthorize": {
      "get": {
        "tags": [
          "Test"
        ],
        "operationId": "TestAuthorize",
        "responses": {
          "200": {
            "description": "Success"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Test/TestAuthenticateAdmin": {
      "get": {
        "tags": [
          "Test"
        ],
        "operationId": "TestAuthenticateAdmin",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Test/TestTime": {
      "get": {
        "tags": [
          "Test"
        ],
        "operationId": "TestTime",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DateTimePoco"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateTimePoco"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateTimePoco"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/Authorize": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Authorize",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/Refresh": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/Register": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/User/Logout": {
      "delete": {
        "tags": [
          "User"
        ],
        "operationId": "Logout",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/User/Unregister": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "Unregister",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/User/InsertShot": {
      "post": {
        "tags": [
          "User"
        ],
        "operationId": "InsertShot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shot"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Shot"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Shot"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BigInteger": {
        "type": "object",
        "properties": {
          "isPowerOfTwo": {
            "type": "boolean",
            "readOnly": true
          },
          "isZero": {
            "type": "boolean",
            "readOnly": true
          },
          "isOne": {
            "type": "boolean",
            "readOnly": true
          },
          "isEven": {
            "type": "boolean",
            "readOnly": true
          },
          "sign": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "BinaryData": {
        "type": "object",
        "additionalProperties": false
      },
      "ByteArrayToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DateTimePoco": {
        "type": "object",
        "properties": {
          "dateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DualToken": {
        "type": "object",
        "properties": {
          "tokenA": {
            "type": "string",
            "nullable": true
          },
          "tokenB": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "HashAndSalt": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "salt": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Password": {
        "type": "object",
        "properties": {
          "rawPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Shot": {
        "type": "object",
        "properties": {
          "user_id": {
            "$ref": "#/components/schemas/BigInteger"
          },
          "frame_id": {
            "$ref": "#/components/schemas/BigInteger"
          },
          "ball_id": {
            "$ref": "#/components/schemas/BigInteger"
          },
          "video_id": {
            "$ref": "#/components/schemas/BigInteger"
          },
          "pins_remaining": {
            "$ref": "#/components/schemas/BinaryData"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "lane_Number": {
            "$ref": "#/components/schemas/BinaryData"
          },
          "ddx": {
            "type": "number",
            "format": "float"
          },
          "ddy": {
            "type": "number",
            "format": "float"
          },
          "ddz": {
            "type": "number",
            "format": "float"
          },
          "dx": {
            "type": "number",
            "format": "float"
          },
          "dy": {
            "type": "number",
            "format": "float"
          },
          "dz": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "UserIdentification": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "nullable": true
          },
          "lastname": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}